refactor(subscriptions): move subscription payment logic to shared utils
Extract handleSubscriptionPayment function from route.ts to a shared utility file for better code organization and reusability across different endpoints
